backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.9.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
google-generativeai==0.3.2
elevenlabs==0.2.27
aiohttp==3.9.1
openai-whisper==20231117
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
praat-parselmouth==0.4.3
silero-vad==5.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
gTTS==2.5.0
